[22] more project card add, for project common coponent created, view port issue resolve
