+++ .env
@@ -3,12 +3,19 @@ REACT_APP_WHATSAPP_URL = https://api.whatsapp.com/send?phone=919689671427
 REACT_APP_GITHUB_URL = https://github.com/atuljadhav93
 REACT_APP_YOUTUBE_URL = https://www.youtube.com/@techedumeet
 REACT_APP_INSTAGRAM_URL = https://instagram.com/atul_jadhav93?utm_source=qr&igshid=MzNlNGNkZWQ4Mg%3D%3D
+
 REACT_APP_MAIL_ID = atuljadhav618@gmail.com
+
 REACT_APP_NETFLIX= https://res.cloudinary.com/dtd0guzj1/image/upload/v1706863487/Portfolio/netflix%20image.png
 REACT_APP_LANCER_LINK = https://res.cloudinary.com/dtd0guzj1/image/upload/v1707934799/Portfolio/Lancer_Link_xzda4o.png
 REACT_APP_PORTFOLIO = https://res.cloudinary.com/dtd0guzj1/image/upload/v1707935116/Portfolio/Atul_Portfolio_vlntab.png
+REACT_APP_MOVIE_WEB = https://res.cloudinary.com/dtd0guzj1/image/upload/v1709727029/Portfolio/MovieWeb_abbw2r.png
+REACT_APP_ADMIN_PORTAL = https://res.cloudinary.com/dtd0guzj1/image/upload/v1709727987/Portfolio/Admin_Portal_jtexlh.png
+REACT_APP_EL = https://res.cloudinary.com/dtd0guzj1/image/upload/v1709729702/Portfolio/Confidential_qknsod.jpg
 
 REACT_APP_LANCER_LINK_URL = https://devlancercorp.b2clogin.com/devlancercorp.onmicrosoft.com/b2c_1_lancercorp_signin_flow/oauth2/v2.0/authorize?response_type=id_token&scope=openid%20profile&client_id=a3952aa6-d239-4981-b1c2-daef94336532&redirect_uri=http%3A%2F%2Flocalhost%3A3000&state=eyJpZCI6IjUxZDFhODkzLTYxODAtNGZmOS1hZjQ2LWI0ZmQ4MTNlMTNjMSIsInRzIjoxNzA3OTM0NTkwLCJtZXRob2QiOiJyZWRpcmVjdEludGVyYWN0aW9uIn0%3D&nonce=9b7e5d74-f88d-458b-8af2-5235761c1f4a&client_info=1&x-client-SKU=MSAL.JS&x-client-Ver=1.4.16&client-request-id=793b0203-90c3-468d-9e5b-6eb072194972&response_mode=fragment
+REACT_APP_EL_ADMIN = https://admin-dev.einstonlabs.com/
+REACT_APP_EL_PORTAL = https://einstonlabs.com/
 
 REACT_APP_RESEUM_LINK = https://collection.cloudinary.com/dtd0guzj1/46cde0ce2e121f30cccc6e5f583df37b
